Moved rendering files into its own folder
