+++ COMP-8551-Project.vcxproj
@@ -146,10 +146,10 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="chunkManagerStandard.cpp" />
-    <ClCompile Include="glad.c" />
+    <ClCompile Include="rendering\glad.c" />
     <ClCompile Include="Main.cpp" />
     <ClCompile Include="protoChunkManager.cpp" />
-    <ClCompile Include="RenderTutorial.cpp" />
+    <ClCompile Include="rendering\RenderTutorial.cpp" />
   </ItemGroup>
   <ItemGroup>
     <None Include=".editorconfig" />
@@ -160,8 +160,13 @@
     <ClInclude Include="chunkManagerStandard.h" />
     <ClInclude Include="entityID.h" />
     <ClInclude Include="protoChunkManager.h" />
+<<<<<<< HEAD
     <ClInclude Include="RenderTutorial.h" />
     <ClInclude Include="sigArch.h" />
+=======
+    <ClInclude Include="rendering\RenderTutorial.h" />
+    <ClInclude Include="SigArch.h" />
+>>>>>>> 416d03f... Moved rendering files into its own folder
     <ClInclude Include="tempPosition.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
